Creation of IniParse gem scaffolding.
IniParse will be a Ruby library for parsing and writing to INI files.
The library will aim to achieve similar goals to the Python library of
the same name.

In particular:
 * Preservation of non-trailing whitespace and blank lines.
 * Support for duplicate options and sections.
 * Preservation of section and option ordering.
 * (maybe-perhaps) Support for multi-line options with backslashes.

Note: I have absolutely no involvement on the Python iniparse; it
simply serves as my inspiration in creating this Ruby version.