Section now acts as a collection by composing OptionCollection.
Section makes #[] and #[]= available for easy access to - and setting
of - Option values. The collection is exposed through #lines, with
Section#each providing an easy way to enumerate through each Option.

Finally, #merge! can be used to slam two Sections together, such as in
cases where two sections with the same name are found in an INI doc.